requirements.txt
langchain
streamlit
langchain-google-genai
python-dotenv
google.generativeai
pandas